Moving the types to a different file
