+++ packages/wdio-browserstack-service/src/@types/bstack-service-types.d.ts
@@ -17,60 +17,3 @@ declare namespace WebdriverIO {
         stopA11yScanning: () => Promise<void>
     }
 }
-
-declare global {
-    interface State {
-        value: number,
-        toString: () => string
-    }
-
-    interface TestContextOptions {
-        skipSessionName: boolean,
-        skipSessionStatus: boolean,
-        sessionNameOmitTestTitle: boolean,
-        sessionNamePrependTopLevelSuiteTitle: boolean,
-        sessionNameFormat: (
-            config: Partial<Options.Testrunner>,
-            capabilities: Partial<Capabilities.RemoteCapability>,
-            suiteTitle: string,
-            testTitle?: string
-        ) => string
-    }
-
-    interface GRRUrls {
-        automate: {
-            hub: string,
-            cdp: string,
-            api: string,
-            upload: string
-        },
-        appAutomate: {
-            hub: string,
-            cdp: string,
-            api: string,
-            upload: string
-        },
-        percy: {
-            api: string
-        },
-        turboScale: {
-            api: string
-        },
-        accessibility: {
-            api: string,
-        },
-        appAccessibility: {
-            api: string
-        },
-        observability: {
-            api: string,
-            upload: string
-        },
-        configServer: {
-            api: string
-        },
-        edsInstrumentation: {
-            api: string
-        }
-    }
-}
