+++ packages/wdio-browserstack-service/tsconfig.json
@@ -8,7 +8,6 @@
   },
   "include": [
     "src/**/*",
-    "../../@types",
-    "src/@types/**/*"
+    "../../@types"
   ]
 }
